Extract Mapper as Single Source of Truth
